Refactor practitioner referencing code
