+++ src/fhir/system-urls.ts
@@ -1,13 +1,25 @@
-export const SYSTEM_CCS =
-  "https://www.hcup-us.ahrq.gov/toolssoftware/ccsr/dxccsr.jsp";
+// HL7 / FHIR
 
+export const SYSTEM_BASIC_RESOURCE_TYPE = `http://terminology.hl7.org/CodeSystem/basic-resource-type`;
 export const SYSTEM_CONDITION_CLINICAL =
   "http://terminology.hl7.org/CodeSystem/condition-clinical";
-
 export const SYSTEM_CONDITION_VERIFICATION_STATUS =
   "http://terminology.hl7.org/CodeSystem/condition-ver-status";
+export const SYSTEM_MARITAL_STATUS =
+  "http://terminology.hl7.org/CodeSystem/v3-MaritalStatus";
+export const SYSTEM_PROVENANCE_ACTIVITY_TYPE =
+  "http://terminology.hl7.org/3.1.0/CodeSystem-v3-DataOperation.html";
+export const SYSTEM_PROVENANCE_AGENT_TYPE =
+  "http://terminology.hl7.org/CodeSystem/provenance-participant-type";
+export const SYSTEM_RELATIONSHIP =
+  "http://terminology.hl7.org/CodeSystem/v2-0131";
+export const SYSTEM_TASK_STATUS =
+  "https://www.hl7.org/fhir/codesystem-task-status.html";
 
-export const SYSTEM_ENRICHMENT = "https://zusapi.com/terminology/enrichment";
+// CODE SYSTEMS
+
+export const SYSTEM_CCS =
+  "https://www.hcup-us.ahrq.gov/toolssoftware/ccsr/dxccsr.jsp";
 
 export const SYSTEM_ICD9 = "http://hl7.org/fhir/sid/icd-9";
 
@@ -17,38 +29,32 @@ export const SYSTEM_ICD10 = "http://hl7.org/fhir/sid/icd-10";
 
 export const SYSTEM_ICD10_CM = "http://hl7.org/fhir/sid/icd-10-cm";
 
-export const SYSTEM_MARITAL_STATUS =
-  "http://terminology.hl7.org/CodeSystem/v3-MaritalStatus";
-
-export const SYSTEM_PRACTITIONER_ID = "https://zusapi.com/practitioner_id";
-
-export const SYSTEM_PROVENANCE_ACTIVITY_TYPE =
-  "http://terminology.hl7.org/3.1.0/CodeSystem-v3-DataOperation.html";
-
-export const SYSTEM_PROVENANCE_AGENT_TYPE =
-  "http://terminology.hl7.org/CodeSystem/provenance-participant-type";
-
-export const SYSTEM_RELATIONSHIP =
-  "http://terminology.hl7.org/CodeSystem/v2-0131";
-
 export const SYSTEM_RXNORM = "http://www.nlm.nih.gov/research/umls/rxnorm";
 
 export const SYSTEM_SNOMED = "http://snomed.info/sct";
 
-export const SYSTEM_SUMMARY = "https://zusapi.com/summary";
+// ZUS
 
-export const SYSTEM_TASK_STATUS =
-  "https://www.hl7.org/fhir/codesystem-task-status.html";
+export const SYSTEM_SUMMARY = "https://zusapi.com/summary";
 
 export const SYSTEM_ZUS_BUILDER_ID = "https://zusapi.com/builder_id";
 export const SYSTEM_ZUS_LENS = "https://zusapi.com/lens";
 export const SYSTEM_ZUS_OWNER = "https://zusapi.com/accesscontrol/owner";
+export const SYSTEM_ZUS_PRACTITIONER_ID = "https://zusapi.com/practitioner_id";
+export const SYSTEM_ZUS_PROFILE_ACTION = "https://zusapi.com/profile/action";
 export const SYSTEM_ZUS_THIRD_PARTY = "https://zusapi.com/thirdparty/source";
 export const SYSTEM_ZUS_UNIVERSAL_ID =
   "https://zusapi.com/fhir/identifier/universal-id";
 export const SYSTEM_ZUS_UPI_RECORD_TYPE =
   "https://zusapi.com/fhir/tag/upi-record-type";
 export const SYSTEM_ZUS_USER_TYPE = "https://zusapi.com/user_type";
+
+// ENRICHMENT
+
+export const SYSTEM_ENRICHMENT = "https://zusapi.com/terminology/enrichment";
+
+// LENS
+
 export const LENS_EXTENSION_MEDICATION_LAST_FILL_DATE =
   "https://zusapi.com/lens/extension/medicationLastFillDate";
 export const LENS_EXTENSION_MEDICATION_LAST_PRESCRIBED_DATE =
@@ -63,5 +69,8 @@ export const LENS_EXTENSION_MEDICATION_LAST_PRESCRIBER =
   "https://zusapi.com/lens/extension/medicationLastPrescriber";
 export const LENS_EXTENSION_AGGREGATED_FROM =
   "https://zusapi.com/lens/extension/aggregatedFrom";
+
+// MISC
+
 export const CTW_EXTENSION_LENS_AGGREGATED_FROM =
   "https://zusapi.com/ctw-component-library/extension/lensAggregatedFrom";
